corrected mentor's remarks
