+++ src/card/card.js
@@ -2,35 +2,13 @@ import {
   createElement
 } from '../utils.js';
 import {
-  TYPES_MAP
-} from '../trip-types.js';
+  getTemplate
+} from './createCardTemplate.js';
 
-
-const createOffers = (specialsArray) => {
-  let specials = ``;
-  for (let special of specialsArray) {
-    specials +=
-      `<li>
-        <button class="trip-point__offer">${special.name} +&euro;&nbsp;${special.price}</button>
-      </li>`;
-  }
-  return specials;
-};
-
-const renderOffersTemplate = (specials) =>
-  `<ul class="trip-point__offers">
-    ${createOffers(specials)}
-  </ul>`;
 
 export default class Card {
   constructor(data) {
-    this._id = data.id;
-    this._type = data.type;
-    this._destination = data.destination;
-    this._time = data.time;
-    this._duration = data.duration;
-    this._price = data.price;
-    this._specials = data.specials;
+    this._data = data;
 
     this._element = null;
     this._onClickInjected = null;
@@ -38,16 +16,7 @@ export default class Card {
   }
 
   get template() {
-    return `<article class="trip-point">
-    <i class="trip-icon">${TYPES_MAP[this._type].icon}</i>
-    <h3 class="trip-point__title">${TYPES_MAP[this._type].title} ${this._destination}</h3>
-    <p class="trip-point__schedule">
-      <span class="trip-point__timetable">${this._time}</span>
-      <span class="trip-point__duration">${this._duration}</span>
-    </p>
-    <p class="trip-point__price">&euro;&nbsp;${this._price}</p>
-    ${renderOffersTemplate(this._specials)}
-  </article>`;
+    return getTemplate(this._data);
   }
 
   get element() {
@@ -55,7 +24,7 @@ export default class Card {
   }
 
   get id() {
-    return this._id;
+    return this._data.id;
   }
 
   bind() {
